move `init` to `google/init.js`
